refactor: Use uvu as test runner
Matches what scip-typescript does; allows passing a flag
