+++ packages/pyright-scip/package.json
@@ -7,8 +7,8 @@
     "build": "webpack --mode production --progress",
     "clean": "shx rm -rf ./dist ./out README.md LICENSE.txt",
     "prepack": "npm run clean && shx cp ../../README.md . && shx cp ../../LICENSE.txt . && npm run build",
-    "check-snapshots-new": "jest check-snapshots",
-    "update-snapshots-new": "jest update-snapshots",
+    "check-snapshots-new": "uvu -r ts-node/register --ignore dist --check",
+    "update-snapshots-new": "uvu -r ts-node/register --ignore dist",
     "check-snapshots": "npm run update-snapshots -- --check",
     "update-snapshots": "node --enable-source-maps ./index.js snapshot-dir snapshots --environment snapshots/testEnv.json --quiet",
     "test": "jest --forceExit --detectOpenHandles",
@@ -41,6 +41,7 @@
     "ts-loader": "^9.2.6",
     "tsconfig-paths-webpack-plugin": "^3.5.2",
     "typescript": "^4.5.4",
+    "uvu": "^0.5.6",
     "webpack": "^5.65.0",
     "webpack-cli": "^4.9.1"
   },
