--- packages/pyright-scip/src/check-snapshots.test.ts
@@ -1,3 +1,0 @@
-import { testMain } from './mainImpl';
-
-testMain('check');
