--- packages/pyright-scip/src/update-snapshots.test.ts
@@ -1,3 +1,0 @@
-import { testMain } from './mainImpl';
-
-testMain('update');
